Moved build_processing_chain out of raw_to_dsp.py and into the dsp dir
